refactored entity classes, added child classes

+++ app/src/main/java/com/example/se306_project1/models/Category.java
@@ -17,7 +17,7 @@ public class Category implements ICategory{
     }
 
     @Exclude
-    public int getCategoryID() {
+    public long getCategoryID() {
         return categoryID;
     }
     @Exclude
